refactor: 경로 수정 및 상수 통합
