+++ src/features/fetchMovies.ts
@@ -1,7 +1,5 @@
 import { Movie } from "../../types/types";
-
-const API_KEY: string = import.meta.env.VITE_TMDB_API_KEY;
-const BASE_URL: string = "https://api.themoviedb.org/3";
+import { API_KEY, BASE_URL } from "../constants/constant";
 
 export async function fetchMovies(
   path: string,
